workers/server/requirements.txt
aiohttp==3.8.1
python-socketio-client==1.1
websocket-client==1.3.1
python-socketio==5.5.2
eventlet==0.33.0
asyncio==3.4.3